apis/shop/serializers.py
from rest_framework import serializers
from .models import Product, Order, OrderItem
# from django.db import transaction

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'

class OrderItemSerializer(serializers.ModelSerializer):
    product = ProductSerializer(read_only=True)

    class Meta:
        model = OrderItem
        fields = ['product', 'quantity']
class OrderSerializer(serializers.ModelSerializer):

    class Meta:
        model = Order
        fields = ['id', 'customer_name', 'email', 'phone_number', 'address', 'status']
        read_only_fields = ['id']
