requirements.txt
djangorestframework==3.14.0
django==5.0.1